Add direction argument to get_ops_cnt function and fix datatypes
- Transition opscount to 2D for real kernels to address forward/backward opscount differences
- Added direction argument to ops_cnt function to support selection if required
- Changed the precision and direction types to UINT8

+++ core/common/bluestein_utils.h
@@ -53,8 +53,8 @@ typedef enum
 INTP get_extended_length(INTP n);
 INT32 prepare_bluestein_sequence(aoclfftz_solution_t *sol, INTP m);
 INT32 elementwise_multiplication(VOID *out, VOID *a, VOID *b, INTP n,
-                                 UINT8 sign, UINT32 precision);
+                                 UINT8 sign, UINT8 precision);
 INT32 normalize_data(VOID *data, INTP n, DOUBLE normalize_factor,
-                     UINT32 precision);
+                     UINT8 precision);
 
 #endif // BLUESTEIN_UTILS_H
